# src/main/java/com/informedeimagenes/db/informedeimagenes_db/entity/Estudios.java
package com.informedeimagenes.db.informedeimagenes_db.entity;

import com.informedeimagenes.db.informedeimagenes_db.entity.base.EstudiosBase;

public class Estudios extends EstudiosBase {

	//OVERRIDE HERE YOUR CUSTOM MAPPER
	
	
}